Revert "Revert "Removed proxy and manager frm std and sdk""
This reverts commit 454acd7433e5de5585f33205db623e315ba18983.
