+++ framework/packages/abstract-sdk/src/apis/bank.rs
@@ -26,8 +26,8 @@ pub trait TransferInterface:
         use abstract_sdk::prelude::*;
         # use cosmwasm_std::testing::mock_dependencies;
         # use abstract_sdk::mock_module::MockModule;
-        # let module = MockModule::new();
         # let deps = mock_dependencies();
+        # let module = MockModule::new(deps.api);
 
         let bank: Bank<MockModule>  = module.bank(deps.as_ref());
         ```
@@ -65,8 +65,8 @@ impl<'a, T: TransferInterface> ApiIdentification for Bank<'a, T> {
     use abstract_sdk::prelude::*;
     # use cosmwasm_std::testing::mock_dependencies;
     # use abstract_sdk::mock_module::MockModule;
-    # let module = MockModule::new();
     # let deps = mock_dependencies();
+    # let module = MockModule::new(deps.api);
 
     let bank: Bank<MockModule>  = module.bank(deps.as_ref());
     ```
@@ -116,6 +116,7 @@ impl<'a, T: TransferInterface + AccountExecutor> Bank<'a, T> {
     /// Transfer the provided funds from the Account to the recipient.
     /// ```
     /// # use cosmwasm_std::{Addr, Response, Deps, DepsMut, MessageInfo};
+    /// # use abstract_std::version_control::Account;
     /// # use abstract_std::objects::AnsAsset;
     /// # use abstract_std::objects::ans_host::AnsHost;
     /// # use abstract_sdk::{
@@ -124,7 +125,7 @@ impl<'a, T: TransferInterface + AccountExecutor> Bank<'a, T> {
     /// # };
     /// # struct MockModule;
     /// # impl AccountIdentification for MockModule {
-    /// #    fn proxy_address(&self, _deps: Deps) -> AbstractSdkResult<Addr> {
+    /// #    fn account(&self, _deps: Deps) -> AbstractSdkResult<Account> {
     /// #       unimplemented!("Not needed for this example")
     /// #   }
     /// # }
@@ -274,7 +275,7 @@ mod test {
     use crate::mock_module::*;
 
     mod transfer_coins {
-        use abstract_std::proxy::ExecuteMsg;
+        use abstract_std::account::ExecuteMsg;
 
         use super::*;
         use crate::{Execution, Executor, ExecutorMsg};
@@ -300,10 +301,10 @@ mod test {
                 amount: coins,
             });
 
-            let base = test_account_base(deps.api);
+            let account = test_account_base(deps.api);
             assert_that!(response.messages[0].msg).is_equal_to(
                 &wasm_execute(
-                    base.proxy,
+                    account.addr(),
                     &ExecuteMsg::ModuleAction {
                         msgs: vec![expected_msg],
                     },
@@ -337,9 +338,9 @@ mod test {
             let response: Response = app.response("deposit").add_messages(deposit_msgs);
             // ANCHOR_END: deposit
 
-            let base = test_account_base(deps.api);
+            let account = test_account_base(deps.api);
             let bank_msg: CosmosMsg = CosmosMsg::Bank(BankMsg::Send {
-                to_address: base.proxy.to_string(),
+                to_address: account.addr().to_string(),
                 amount: coins,
             });
