+++ framework/packages/abstract-sdk/src/base/features/identification.rs
@@ -1,19 +1,14 @@
-use abstract_std::{
-    objects::common_namespace::ADMIN_NAMESPACE, proxy::state::ACCOUNT_ID, version_control::Account,
-};
-use cosmwasm_std::{Addr, Deps};
-use cw_storage_plus::Item;
+use abstract_std::{account::state::ACCOUNT_ID, version_control::Account};
+use cosmwasm_std::Deps;
 
 use crate::std::objects::AccountId;
 // see std::proxy::state::ADMIN
-use crate::{AbstractSdkError, AbstractSdkResult};
-
-const MANAGER: Item<Option<Addr>> = Item::new(ADMIN_NAMESPACE);
+use crate::AbstractSdkResult;
 
 /// Retrieve identifying information about an Account.
 /// This includes the manager, proxy, core and account_id.
 pub trait AccountIdentification: Sized {
-    /// Get the proxy address for the current account.
+    /// Get the account address
     fn account(&self, deps: Deps) -> AbstractSdkResult<Account>;
 
     /// Get the Account id for the current account.
@@ -38,9 +33,9 @@ mod test {
     }
 
     impl AccountIdentification for MockBinding {
-        fn proxy_address(&self, _deps: Deps) -> AbstractSdkResult<Addr> {
+        fn account(&self, _deps: Deps) -> AbstractSdkResult<Account> {
             let account_base = test_account_base(self.mock_api);
-            Ok(account_base.proxy)
+            Ok(account_base)
         }
     }
 
@@ -50,52 +45,14 @@ mod test {
         use super::*;
 
         #[test]
-        fn test_proxy_address() {
+        fn test_account_address() {
             let deps = mock_dependencies();
             let binding = MockBinding { mock_api: deps.api };
 
             let account_base = test_account_base(deps.api);
 
-            let res = binding.proxy_address(deps.as_ref());
-            assert_that!(res).is_ok().is_equal_to(account_base.proxy);
-        }
-
-        #[test]
-        fn test_manager_address() {
-            let mut deps = mock_dependencies();
-            let binding = MockBinding { mock_api: deps.api };
-            let account_base = test_account_base(deps.api);
-
-            deps.querier = MockQuerierBuilder::default()
-                .with_contract_item(
-                    &account_base.proxy,
-                    MANAGER,
-                    &Some(account_base.manager.clone()),
-                )
-                .build();
-
-            assert_that!(binding.manager_address(deps.as_ref()))
-                .is_ok()
-                .is_equal_to(account_base.manager);
-        }
-
-        #[test]
-        fn test_account() {
-            let mut deps = mock_dependencies();
-            let expected_account_base = test_account_base(deps.api);
-
-            deps.querier = MockQuerierBuilder::default()
-                .with_contract_item(
-                    &expected_account_base.proxy,
-                    MANAGER,
-                    &Some(expected_account_base.manager.clone()),
-                )
-                .build();
-
-            let binding = MockBinding { mock_api: deps.api };
-            assert_that!(binding.account_base(deps.as_ref()))
-                .is_ok()
-                .is_equal_to(expected_account_base);
+            let res = binding.account(deps.as_ref());
+            assert_that!(res).is_ok().is_equal_to(account_base);
         }
 
         #[test]
@@ -104,7 +61,7 @@ mod test {
             let account_base = test_account_base(deps.api);
 
             deps.querier = MockQuerierBuilder::default()
-                .with_contract_item(&account_base.proxy, ACCOUNT_ID, &TEST_ACCOUNT_ID)
+                .with_contract_item(account_base.addr(), ACCOUNT_ID, &TEST_ACCOUNT_ID)
                 .build();
 
             let binding = MockBinding { mock_api: deps.api };
