+++ framework/packages/abstract-std/src/account/mod.rs
@@ -1,0 +1,7 @@
+pub mod msgs;
+pub mod responses;
+pub mod state;
+pub mod types;
+
+pub use types::ModuleInstallConfig;
+pub use msgs::{ExecuteMsg, ExecuteMsgFns, InstantiateMsg, MigrateMsg, QueryMsg, QueryMsgFns};
