--- framework/packages/abstract-std/src/core/mod.rs
@@ -1,3 +1,0 @@
-pub mod account;
-pub mod manager;
-pub mod proxy;
