+++ framework/packages/abstract-std/src/standalone.rs
@@ -9,8 +9,6 @@ use crate::{
     version_control::Account,
 };
 
-use cosmwasm_std::Addr;
-
 /// Data required for the `StandaloneContract::instantiate` function.
 #[cosmwasm_schema::cw_serde]
 pub struct StandaloneInstantiateMsg {
